Moving structs about

+++ server/pkg/assets/service.go
@@ -19,7 +19,6 @@ import (
 	"github.com/freshteapot/learnalist-api/server/pkg/openapi"
 	guuid "github.com/google/uuid"
 	"github.com/labstack/echo/v4"
-	"github.com/nats-io/stan.go"
 	"github.com/sirupsen/logrus"
 )
 
@@ -32,9 +31,12 @@ func NewService(directory string, acl acl.AclAsset, repo Repository, logEntry *l
 		repo:      repo,
 	}
 
-	event.GetBus().Subscribe(event.ApiUserDelete, func(msg *stan.Msg) {
-		fmt.Printf("Listening to %s\n", event.ApiUserDelete)
-		userUUID := string(msg.Data)
+	event.GetBus().Subscribe(event.TopicMonolog, func(entry event.Eventlog) {
+		if entry.Kind != event.ApiUserDelete {
+			return
+		}
+
+		userUUID := entry.Data.(string)
 		fmt.Println("event.ApiUserDelete", userUUID)
 		s.DeleteUserAssets(userUUID)
 	})
